add basic serverless stuff
