+++ package.json
@@ -40,6 +40,7 @@
     "nodemon": "^2.0.3",
     "python-shell": "^2.0.1",
     "request": "^2.88.2",
+    "serverless-http": "^2.6.0",
     "snoowrap": "^1.21.0",
     "twitter": "^1.7.1",
     "url": "^0.11.0",
