+++ serverless.yml
@@ -1,0 +1,19 @@
+service: rona-api
+provider:
+  name: aws
+  runtime: nodejs12.x
+  stage: dev
+  region: us-west-1
+  memorySize: 128
+functions:
+  api:
+    handler: handler.api
+    events: 
+      - http: 
+          path: /
+          method: ANY
+          cors: true
+      - http: 
+          path: /{proxy+}
+          method: ANY
+          cors: true
